+++ src/main/kotlin/orm/tableInheritance/ITableInheritanceMapper.kt
@@ -4,9 +4,9 @@ import kotlin.reflect.KClass
 
 interface ITableInheritanceMapper {
     fun insert(entity: Any): Boolean
-    fun find(id: Long? = null): Any?
+    fun find(id: Int? = null): Any?
     fun update(entity: Any): Boolean
-    fun remove(id: Long): Boolean
+    fun remove(id: Int): Boolean
     fun query(q: String): Any? // for hardcore SQL queries like using WHERE, ORDER BY etc. - who needs them btw?
-    fun findWithoutRelations(id: Long, entityClass: KClass<*>): Any?
+    fun findWithoutRelations(id: Int, entityClass: KClass<*>): Any?
 }
